prediction_models/requirements.txt
Flask
flask-cors
eventlet
python-socketio
requests
Pillow
numpy
opencv-python
tensorflow
keras
scikit-learn
mtcnn
insightface
onnxruntime
mediapipe

